# api/openapi.yaml
openapi: 3.1.0
info:
  title: GoSite Panel API
  version: 1.0.0
  description: |
    REST API for the GoSite hosting control panel.

    Authentication uses an HTTP-only session cookie (`gosite_session`) issued by
    `POST /auth/login`. When `AUTH_ENABLE=true`, all `/api/v1/*` routes also
    require HTTP Basic authentication before session checks.

    Errors use a consistent envelope: `{ "error": { "code": "...", "message": "..." } }`.
servers:
  - url: https://localhost:8080/api/v1
    description: Panel API (v1)
  - url: https://localhost:8080
    description: Root (health check only)
...
paths:
...
  /files:
    get:
      tags: [Files]
      summary: Browse directory
      operationId: browseFiles
      security:
        - basicAuth: []
          sessionCookie: []
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
          description: Absolute path; the file manager can browse from filesystem root when permitted by the host process.
      responses:
        "200":
          description: Directory listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Files]
      summary: Create file or directory
      description: |
        Creates a file or directory from JSON, or uploads a file via
        `multipart/form-data` with fields `path` and `file`.
      operationId: createFile
      security:
        - basicAuth: []
          sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileCreateRequest"
          multipart/form-data:
            schema:
              type: object
              required: [path, file]
              properties:
                path:
                  type: string
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Created or uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [Files]
      summary: Delete file or directory
      operationId: deleteFile
      security:
        - basicAuth: []
          sessionCookie: []
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /query/tail:
    get:
      tags: [Observability]
      summary: Live tail of query events
      description: |
        Server-Sent Events stream of `QueryEvent` objects. The connection stays
        open and emits new audit, job, access, and error events as they are
        ingested. Use `event: error` SSE events to observe errors and `: keepalive`
        comments as heartbeats.
      operationId: getQueryTail
      security:
        - basicAuth: []
          sessionCookie: []
      parameters:
        - name: source
          in: query
          schema:
            type: string
            enum: [audit, job, access, error, all]
          description: Source to subscribe to. Defaults to `all`.
        - name: site
          in: query
          schema:
            type: string
          description: Optional nginx site/domain scope for access/error sources.
      responses:
        "200":
          description: SSE stream of QueryEvent frames
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/QueryTailResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: gosite_session
      description: HTTP-only session cookie issued by POST /auth/login
    basicAuth:
      type: http
      scheme: basic
      description: Required on all /api/v1 routes when AUTH_ENABLE=true
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code
              examples:
                - VALIDATION_FAILED
                - AUTH_INVALID_CREDENTIALS
                - SESSION_EXPIRED
                - BASIC_AUTH_REQUIRED
                - NGINX_TEST_FAILED
            message:
              type: string
              description: Human-readable error message
...
    QueryEvent:
      type: object
      required: [ts, source, action, user, message, meta]
      properties:
        ts:
          type: string
          format: date-time
        source:
          type: string
        action:
          type: string
        user:
          type: string
        message:
          type: string
        meta:
          type: object
          additionalProperties: true
...
    FileEntry:
      type: object
      required: [name, path, size, mode, is_dir, mod_time, kind, mime_type, extension, editable, viewable, archive, symlink]
      properties:
        name:
          type: string
        path:
          type: string
        size:
          type: integer
          format: int64
        mode:
          type: string
        is_dir:
          type: boolean
        mod_time:
          type: string
        kind:
          type: string
          enum: [directory, image, json, yaml, config, text, archive, binary]
        mime_type:
          type: string
        extension:
          type: string
        editable:
          type: boolean
        viewable:
          type: boolean
        archive:
          type: boolean
        symlink:
          type: boolean
        target:
          type: string
    FileArchiveTools:
      type: object
      required: [unzip, tar, gzip]
      properties:
        unzip:
          type: boolean
        tar:
          type: boolean
        gzip:
          type: boolean
    FileListResponse:
      type: object
      required: [entries, tools]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/FileEntry"
        tools:
          $ref: "#/components/schemas/FileArchiveTools"
...
    FileCreateRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [directory, file, remote]
        name:
          type: string
        path:
          type: string
        content:
          type: string
...
    QueryTailResponse:
      type: object
      description: |
        Documentation shape for an SSE `/query/tail` stream. Each
        `data: <json>` line in the response carries one `QueryEvent` from
        this envelope. The envelope also accepts an `event: error` SSE event
        whose `data` is an `ApiError` envelope.
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/QueryEvent"
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string